+++ src/com/game/Window.java
@@ -10,13 +10,13 @@ public class Window extends Canvas{                             // class to init
     @Serial
     private static final long serialVersionUID = 5070093916032438121L;
 
-    public Window(int width, int height, String title, Game game) { // window constructor sets parameters
+    public Window(float width, float height, String title, Game game) { // window constructor sets parameters
 
         JFrame frame = new JFrame(title);
 
-        frame.setPreferredSize(new Dimension(width,height));
-        frame.setMaximumSize(new Dimension(width,height));
-        frame.setMinimumSize(new Dimension(width,height));
+        frame.setPreferredSize(new Dimension((int)width,(int)height));
+        frame.setMaximumSize(new Dimension((int)width,(int)height));
+        frame.setMinimumSize(new Dimension((int)width,(int)height));
 
         frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);   // enables closing of game thread window with X button
         frame.setResizable(false);
